requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
zipp==3.21.0

rembg
Pillow
opencv-python-headless==4.10.0.82
numpy==1.26.4
